Edit Category to include new settings
- Move categories funcs to routes_categories.py
- Add new update_categories()
- CRUD for categories fully implemented
