+++ TODO.md
@@ -18,10 +18,11 @@ Here is my list of todoes for the project.
   - [x] update models to move the duration and count
   - [x] migrate the category functions from routes_main.py to routes_category.py
   - [x] update import/export/delete Questions for new category settings
-  - [ ] enhance Edit Category to include the timer_duration and questions_per_quiz
-  - [ ] Make timer and count editable on existing category
-  - [ ] Add a save button that only activates if a field changes
-  - [ ] 
+  - [x] enhance Edit Category to include the timer_duration and questions_per_quiz
+  - [x] Make timer and count editable on existing category
+  - [ ] Add a save button that only activates if a field changes for Edit Category
+  - [ ] Add a header for the Existing Categories grid
+  - [ ] TBD
 
 Migrating from top level to Quiz Category is mid-cycle. So this is not ready for a commit yet.
