+++ templates/edit_categories.html
@@ -9,10 +9,15 @@
     <h2 class="mt-4">Existing Categories</h2>
     <ul class="list-group">
         {% for category in categories %}
-            <li class="list-group-item d-flex justify-content-between align-items-center">
-                {{ category.name }}
-                <form action="/delete_category/{{ category.id }}" method="POST" style="margin: 0;">
-                    <button type="submit" class="btn btn-danger">Delete</button>
+            <li class="list-group-item">
+                <form action="/update_category/{{ category.id }}" method="POST" class="d-flex align-items-center">
+                    <input type="text" name="name" value="{{ category.name }}" class="form-control mr-2" required>
+                    <input type="number" name="timer_duration" value="{{ category.timer_duration }}" class="form-control mr-2" required placeholder="Timer Duration (seconds)">
+                    <input type="number" name="questions_per_quiz" value="{{ category.questions_per_quiz }}" class="form-control mr-2" required placeholder="Questions per Quiz">
+                    <button type="submit" class="btn btn-success mr-2">Save</button>
+                    <form action="/delete_category/{{ category.id }}" method="POST" style="margin: 0;">
+                        <button type="submit" class="btn btn-danger">Delete</button>
+                    </form>
                 </form>
             </li>
         {% endfor %}
@@ -23,6 +28,14 @@
         <div class="form-group">
             <label for="category_name">Category Name:</label>
             <input type="text" id="category_name" name="category_name" required class="form-control">
+        </div>
+        <div class="form-group">
+            <label for="timer_duration">Timer Duration (seconds):</label>
+            <input type="number" id="timer_duration" name="timer_duration" required class="form-control" placeholder="Timer Duration (seconds)">
+        </div>
+        <div class="form-group">
+            <label for="questions_per_quiz">Questions per Quiz:</label>
+            <input type="number" id="questions_per_quiz" name="questions_per_quiz" required class="form-control" placeholder="Questions per Quiz">
         </div>
         <button type="submit" class="btn btn-success mt-3">Add Category</button>
     </form>
